+++ NEWS.rst
@@ -6,7 +6,14 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Renamed
+  ``storage_size`` to ``size``,
+  ``storage_cast`` to ``cast``,
+  ``storage_copy`` to ``copy``,
+  ``zero_storage`` to ``fill_zero``,
+  ``split_storage`` to ``split``.
+* Major: Changed the structure of the repository to be one file per definition.
+* Major: Changed the ``storage`` use flag to ``storage_includes``.
 
 1.0.0
 -----
