+++ README.rst
@@ -2,9 +2,10 @@
 storage
 =======
 
-storage is a C++ header only library providing an abstraction for storage / buffers.
+storage is a C++ header only library providing an abstraction for storage /
+buffers.
 
 License
-=====
+=======
 
 storage is available under the BSD license.
