+++ buildbot.py
@@ -1,5 +1,5 @@
 ﻿#!/usr/bin/env python
-# encodingsak: utf-8
+# encoding: utf-8
 
 import os
 import sys
