--- src/storage/wscript_build
@@ -1,9 +1,0 @@
-# encoding: utf-8
-
-bld.stlib(
-    features='cxx',
-    source=bld.path.ant_glob('**/*.cpp'),
-    target='storage',
-    use=[],
-    export_includes=['..']
-)
