+++ test/src/test_cast.cpp
@@ -1,0 +1,6 @@
+// Copyright (c) Steinwurf ApS 2016.
+// All Rights Reserved
+//
+// Distributed under the "BSD License". See the accompanying LICENSE.rst file.
+
+#include <storage/cast.hpp>
